verteilerschrank updated

+++ WindowMasterConfig.h
@@ -79,10 +79,10 @@ extern const short functionPin[ARRAYSIZE];
 extern const short statePin[ARRAYSIZE];
 extern const char * windowName[ARRAYSIZE];
 
-extern long lastReadCurrent[3];
+extern float lastReadCurrent[3];
 extern const char* currentGA[3];
-extern long current[3];
-extern long uRef;
+extern float current[3];
+extern float uRef;
 
 extern short windowState[ARRAYSIZE];
